Define signals for all actions
Take screenshot is the most sophistaicated. We can now take screenshot of a window using the astuce the focus. When we click on another window, we can emit the focus lost signal in gtkmm-4.0 with EventControllerFocus::signal_leave(). We can then use `gnome-screnshot -w`to take the screenshot of the window that got the focus. Very interesting! :-) open source >> all!